Fix authentication event order and cross-platform acceptance blockers
Publish Authenticated before policy reevaluation can emit Approved. Expose the existing departure facade only to transport-lab integrations, promote the mode_t openat variadic argument for macOS, and use canonical public identity in the opaque daemon fixture.

Workspace strict Clippy and formatting passed. Independent source verification and audit precede candidate publication; updated-head Linux/TURN/macOS and release acceptance remain required. PR remains draft/HOLD.

+++ crates/myownmesh-core/src/persist.rs
@@ -164,7 +164,7 @@ impl DirectoryCapability {
                 self.fd.as_raw_fd(),
                 name.as_ptr(),
                 libc::O_WRONLY | libc::O_CREAT | libc::O_EXCL | libc::O_CLOEXEC | libc::O_NOFOLLOW,
-                mode as libc::mode_t,
+                mode as libc::mode_t as libc::c_uint,
             )
         };
         if raw < 0 {
